wip: cleanup
