+++ app/nexus/internal/initialization/recovery/shamir.go
@@ -2,7 +2,7 @@
 //  \\\\\ Copyright 2024-present SPIKE contributors.
 // \\\\\\\ SPDX-License-Identifier: Apache-2.0
 
-package initialization
+package recovery
 
 import (
 	"sort"
